third_party/blink/web_tests/fast/xsl/xslt/extra-lf-at-end-quirk-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: XSLTProcessor is not defined
Test for bug 15919: XSLTProcessor.transformToFragment creates an extra not defined TEXT_NODE at the end.
